html/cross-origin-embedder-policy/none.tentative.html

<!doctype html>
<meta name="timeout" content="long">
<title>Cross-Origin header and nested navigable resource without such header</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/get-host-info.sub.js"></script>
<div id=log></div>
<script>
async_test(t => {
  const frame = document.createElement("iframe");
  frame.onload = t.step_func_done(() => {
    assert_not_equals(frame.contentDocument, null);
  });
  frame.src = "/common/blank.html";
  document.body.append(frame);
  assert_equals(frame.contentDocument.body.localName, "body");
}, `"none" top-level: navigating a frame to "none" should succeed`);

async_test(t => {
  const frame = document.createElement("iframe");
  let firstNavOk = false;
  frame.onload = t.step_func(() => {
    assert_not_equals(frame.contentDocument, null);
    firstNavOk = true;
  });
  t.step_timeout(() => {
    assert_equals(firstNavOk, true);
    assert_not_equals(frame.contentDocument, null);
    t.done();
  }, 500);
  frame.src = "resources/navigate-require-corp.sub.html?to=/common/blank.html";
  document.body.append(frame);
  assert_equals(frame.contentDocument.body.localName, "body");
}, `"none" top-level: navigating a frame from "require-corp" to "none" should succeed`);

async_test(t => {
  let w = window.open(`resources/navigate-none.sub.html?to=navigate-require-corp.sub.html`, "window_name");

  t.add_cleanup(() => w.close());

  t.step_timeout(() => {
    w.history.back();
    t.step_timeout(() => {
      assert_not_equals(w.document, null);
      t.done();
    }, 500);
  }, 500);
}, `"none" top-level: navigating a frame back from "require-corp" should succeed`);

async_test(t => {
  let pageLoaded = false;
  const CHANNEL_NAME = "require-corp-none-top-navigation";
  let bc = new BroadcastChannel(CHANNEL_NAME);
  let finished = false;
  bc.onmessage = t.step_func((event) => {
    pageLoaded = true;
    let payload = event.data;
    assert_equals(payload, "loaded");
  });

  const SECOND_CHANNEL = "require-corp-none-top-navigation-final";
  let bc2 = new BroadcastChannel(SECOND_CHANNEL);
  bc2.onmessage = t.step_func((event) => {
    finished = true;
    let payload = event.data;
    assert_equals(payload, "loaded");
  });

  let win = window.open(`resources/navigate-require-corp.sub.html?channelName=${CHANNEL_NAME}&to=navigate-none.sub.html?channelName=${SECOND_CHANNEL}`, "_blank", "noopener");

  t.step_timeout(() => {
    assert_equals(pageLoaded, true);
    assert_equals(finished, true);
    t.done();
  }, 500);
}, `"require-corp" top-level noopener popup: navigating to "none" should succeed`);
</script>
